Add a SendConnData method and use payload length headers for unix
